requirements.txt
Flask==2.3.2
psutil==5.9.5
boto3==1.35.0
python-dotenv==1.0.1
